refactor: 리렌더링 로직을 한 곳으로 옮김
